working to clean up tests
