+++ .eslintrc
@@ -9,6 +9,7 @@
   "env": {
     "browser": true,
     "node": true,
+    "jest": true,
     "es6" : true
   },
   "parser": "babel-eslint",
