+++ App.test.js
@@ -6,12 +6,16 @@ import App from "./App";
 
 describe("<App />", () => {
     it("has 1 child", () => {
-        const tree = renderer.create(<App />).toJSON();
+        const tree = renderer.create(
+            <App />
+        ).toJSON();
         expect(tree.children.length).toBe(1);
     });
 
     it("renders correctly", () => {
-        const tree = renderer.create(<App />).toJSON();
+        const tree = renderer.create(
+            <App />
+        ).toJSON();
         expect(tree).toMatchSnapshot();
     });
 });
