+++ __tests__/App-test.js
@@ -1,22 +1,22 @@
-import 'react-native';
-import React from 'react';
-import App from '../App';
-import renderer from 'react-test-renderer';
-import NavigationTestUtils from 'react-navigation/NavigationTestUtils';
+import "react-native";
+import React from "react";
+import App from "../App";
+import renderer from "react-test-renderer";
+import NavigationTestUtils from "react-navigation/NavigationTestUtils";
 
 describe('App snapshot', () => {
-  jest.useFakeTimers();
-  beforeEach(() => {
-    NavigationTestUtils.resetInternalState();
-  });
+    jest.useFakeTimers();
+    beforeEach(() => {
+        NavigationTestUtils.resetInternalState();
+    });
 
-  it('renders the loading screen', async () => {
-    const tree = renderer.create(<App />).toJSON();
-    expect(tree).toMatchSnapshot();
-  });
+    it('renders the loading screen', async () => {
+        const tree = renderer.create(<App />).toJSON();
+        expect(tree).toMatchSnapshot();
+    });
 
-  it('renders the root without loading screen', async () => {
-    const tree = renderer.create(<App skipLoadingScreen />).toJSON();
-    expect(tree).toMatchSnapshot();
-  });
+    it('renders the root without loading screen', async () => {
+        const tree = renderer.create(<App skipLoadingScreen />).toJSON();
+        expect(tree).toMatchSnapshot();
+    });
 });
