+++ components/mono-text/mono-text.js
@@ -1,11 +1,10 @@
-/* global it jest describe expect */
-import "react-native";
 import React from "react";
-import MonoText from "./index";
-import renderer from "react-test-renderer";
+import { Text } from "react-native";
 
-it("renders correctly", () => {
-    const tree = renderer.create(<MonoText>Snapshot test!</MonoText>).toJSON();
+type Props = { style: Object }
 
-    expect(tree).toMatchSnapshot();
-});
+export const MonoText = (props: Props) => {
+    const { style, ...passThroughProps } = props;
+    return (<Text {...passThroughProps} style={[style, { fontFamily: "space-mono" }]}/>);
+};
+
